cleanup lib.rs: create config and ip modules
