+++ src/config.rs
@@ -1,0 +1,71 @@
+use {parse, Ip, Network, ParseError};
+
+/// Encompasses the nameserver configuration
+///
+/// Currently the options and defaults match those of linux/glibc
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct Config {
+    /// List of nameservers
+    pub nameservers: Vec<Ip>,
+    /// List of suffixes to append to name when it doesn't contain ndots
+    pub search: Vec<String>,
+    /// List of preferred addresses
+    pub sortlist: Vec<Network>,
+    /// Enable DNS resolve debugging
+    pub debug: bool,
+    /// Number of dots in name to try absolute resolving first (default 1)
+    pub ndots: u32,
+    /// Dns query timeout (default 5 [sec])
+    pub timeout: u32,
+    /// Number of attempts to resolve name if server is inaccesible (default 2)
+    pub attempts: u32,
+    /// Round-robin selection of servers (default false)
+    pub rotate: bool,
+    /// Don't check names for validity (default false)
+    pub no_check_names: bool,
+    /// Try AAAA query before A
+    pub inet6: bool,
+    /// Use reverse lookup of ipv6 using bit-label format described instead
+    /// of nibble format
+    pub ip6_bytestring: bool,
+    /// Do ipv6 reverse lookups in ip6.int zone instead of ip6.arpa
+    /// (default false)
+    pub ip6_dotint: bool,
+    /// Enable dns extensions described in RFC 2671
+    pub edns0: bool,
+    /// Don't make ipv4 and ipv6 requests simultaneously
+    pub single_request: bool,
+    /// Use same socket for the A and AAAA requests
+    pub single_request_reopen: bool,
+    /// Don't resolve unqualified name as top level domain
+    pub no_tld_query: bool,
+    /// Force using TCP for DNS resolution
+    pub use_vc: bool,
+}
+
+impl Config {
+    pub fn new() -> Config {
+        Config {
+            nameservers: Vec::new(),
+            search: Vec::new(),
+            sortlist: Vec::new(),
+            debug: false,
+            ndots: 1,
+            timeout: 5,
+            attempts: 2,
+            rotate: false,
+            no_check_names: false,
+            inet6: false,
+            ip6_bytestring: false,
+            ip6_dotint: false,
+            edns0: false,
+            single_request: false,
+            single_request_reopen: false,
+            no_tld_query: false,
+            use_vc: false,
+        }
+    }
+    pub fn parse<T: AsRef<[u8]>>(buf: T) -> Result<Config, ParseError> {
+        parse(buf.as_ref())
+    }
+}
